+++ src/main/java/bowling/model/score/FinalScore.java
@@ -1,0 +1,18 @@
+package bowling.model.score;
+
+import java.util.List;
+
+public class FinalScore implements Score {
+
+    private List<Integer> scores;
+
+    public FinalScore(List<Integer> scores) {
+        this.scores = scores;
+    }
+
+    @Override
+    public int currentScore() {
+        return 0;
+    }
+
+}
