+++ src/main/java/bowling/model/score/MissScore.java
@@ -1,4 +1,4 @@
-package bowling.model;
+package bowling.model.score;
 
 public class MissScore implements Score {
 
@@ -17,13 +17,5 @@ public class MissScore implements Score {
 	@Override
 	public int currentScore() {
 		return first + second;
-	}
-
-	@Override
-	public String toString() {
-		if (second == -1) {
-			return String.format(" %s ", this.first == 0 ? "-" : this.first);
-		}
-		return String.format(" %s|%s ", (this.first == 0) ? "-" : this.first, (this.second == 0) ? "-" : this.second);
 	}
 }
