+++ src/main/java/bowling/model/score/SpareScore.java
@@ -1,4 +1,4 @@
-package bowling.model;
+package bowling.model.score;
 
 public class SpareScore implements Score {
 
@@ -11,11 +11,6 @@ public class SpareScore implements Score {
 	@Override
 	public int currentScore() {
 		return 10;
-	}
-
-	@Override
-	public String toString() {
-		return String.format(" %s|/ ", this.score == 0 ? "-" : this.score);
 	}
 
 }
