+++ src/main/java/bowling/model/score/StrikeScore.java
@@ -1,4 +1,4 @@
-package bowling.model;
+package bowling.model.score;
 
 public class StrikeScore implements Score {
 
@@ -8,11 +8,6 @@ public class StrikeScore implements Score {
 	@Override
 	public int currentScore() {
 		return 10;
-	}
-
-	@Override
-	public String toString() {
-		return String.format("  X  ");
 	}
 
 }
